requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-rest-auth==6.0.0
Django
django-allauth
djangorestframework
djangorestframework-simplejwt
et-xmlfile
filelock==3.15.4
fsspec==2024.6.1
graphene==3.4
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
idna==3.10
mysql-connector-python==8.0.31
neo4j==5.25.0
neo4j-driver==5.25.0
pillow==10.4.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.1.1
requests==2.32.3
gunicorn==20.1.0
psycopg2-binary==2.9.3
dj-database-url==0.5.0
celery[redis]
markdown
django-filter